requirements.txt
flask
docx2txt
pypdf
langchain
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
ctransformers
openai
python-dotenv